+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # WIP
 
-- Support heterogeneous (`:db/tupleTypes`) and homogeneous (`:db/tupleType`) tuples #364
+- Support heterogeneous (`:db/tupleTypes`) and homogeneous (`:db/tupleType`) tuples #364 #427
 
 # 1.7.8 - Oct 11, 2025